Added ability to specify colours
