+++ HoloPipelines/core/adapters/glb_file.py
@@ -1,9 +1,45 @@
 """
-This module contains functionality related to transforming OBJ to GLB files.
+This module contains functionality related to writing a mesh to disk as an OBJ file.
 """
+import numpy as np
+import math
 
-from core.wrappers.obj2gltf import call_obj2gltf
+import trimesh
+from trimesh import visual
 
 
-def convert_obj_to_glb_and_write(input_obj_path: str, output_glb_path: str) -> None:
-    call_obj2gltf(input_obj_path, output_glb_path)
+import logging
+
+def write_mesh_as_glb(
+    meshes, output_obj_file_path: str, metadata={}
+) -> None:
+    scene = trimesh.Scene(metadata=metadata)
+    for mesh in meshes:
+        mesh2 = trimesh.Trimesh(vertices=mesh[0],
+                           faces=mesh[1],
+                           vertex_normals=mesh[2])
+
+        mesh2.apply_transform(trimesh.transformations.rotation_matrix(90,(0,0,1)))
+        scene.add_geometry(mesh2)
+    scene.export(output_obj_file_path)
+
+
+def write_mesh_as_glb(
+        meshes, output_obj_file_path: str,colour, metadata={}
+) -> None:
+    scene = trimesh.Scene(metadata=metadata)
+    index = 0
+    for mesh in meshes:
+        mesh2 = trimesh.Trimesh(vertices=mesh[0],
+                                faces=mesh[1],
+                                vertex_normals=mesh[2])
+
+        mesh2.visual.material = trimesh.visual.material.SimpleMaterial(
+            diffuse=np.asarray(colour[index]))
+        print(colour[index])
+        mesh2.apply_transform(trimesh.transformations.rotation_matrix(90, (0, 0, 1)))
+        scene.add_geometry(mesh2)
+        index += 1
+    scene.export(output_obj_file_path)
+
+
