+++ HoloPipelines/core/pipelines/bone_segmentation.py
@@ -9,8 +9,8 @@ import sys
 import numpy as np
 
 from core.adapters.dicom_file import read_dicom_as_np_ndarray_and_normalise
-from core.adapters.glb_file import convert_obj_to_glb_and_write
-from core.adapters.obj_file import write_mesh_as_glb
+
+from core.adapters.glb_file import write_mesh_as_glb
 from core.services.marching_cubes import generate_mesh
 from core.services.np_image_manipulation import downscale_and_conditionally_crop
 from core.tasks.shared.dispatch_output import dispatch_output
