+++ HoloPipelines/core/pipelines/view_nifti.py
@@ -10,13 +10,13 @@ Xinyang Feng, Jie Yang, Andrew F. Laine, Elsa D. Angelini
 import os
 
 from core.adapters.dicom_file import read_dicom_as_np_ndarray_and_normalise
-from core.adapters.glb_file import convert_obj_to_glb_and_write
+
 from core.adapters.nifti_file import (
     convert_dicom_np_ndarray_to_nifti_image,
     read_nifti_as_np_array,
     write_np_array_as_nifti_image,
 )
-from core.adapters.obj_file import write_mesh_as_glb
+from core.adapters.glb_file import write_mesh_as_glb
 from core.services.marching_cubes import generate_mesh
 from core.services.np_image_manipulation import downscale_and_conditionally_crop
 from core.tasks.shared.dispatch_output import dispatch_output
@@ -54,7 +54,7 @@ def run(job_id: str, input_endpoint: str, medical_data: dict) -> None:
     segment.append(generate_mesh(nifti_image_as_np_array5, hu_threshold))
     segment.append(generate_mesh(nifti_image_as_np_array1, hu_threshold))
 
-    write_mesh_as_glb(segment, obj_output_path,{
+    write_mesh_as_glb(segment, obj_output_path,[[0,0.3,1.0,0.2],[1.0,1.0,0.0,1.0]],{
     1: "cortical_gray_matter",
     2: "basal_ganglia",
 })
